+++ storybook/storybook/components/simple_form.story.exs
@@ -38,6 +38,30 @@ defmodule Storybook.Components.SimpleForm do
           type: "select",
           options: ["Option 1", "Option 2", "Option 3"]
         }
+      },
+      %Variation{
+        id: :checkbox_group,
+        attributes: %{
+          label: "Checkbox Group",
+          type: "checkbox_group",
+          options: [
+            {"Option 1", "option1"},
+            {"Option 2", "option2"},
+            {"Option 3", "option3"}
+          ]
+        }
+      },
+      %Variation{
+        id: :radio_group,
+        attributes: %{
+          label: "Radio Group",
+          type: "radio_group",
+          options: [
+            {"Option 1", "option1"},
+            {"Option 2", "option2"},
+            {"Option 3", "option3"}
+          ]
+        }
       }
     ]
   end